Update helper.py
Trying to run `generate_mesh` on a one-dimensional geometry raises a `KeyError` because in one-dimension `cell_data` won't have a 'triangle' key.  Lloyd smoothing is unnecessary in this case anyway, so just return the unsmoothed mesh.

One could test `'triangle' in cell_data` but the PR is consistent with the current floating-point test for three-dimensionality.