Implement ellipsoids for the OpenCascade engine
This adds basic capabilities to add ellipsoids with the OpenCascade
engine. The actual implementation idea is from this mailing list post
by Christophe Geuzaine:
http://onelab.info/pipermail/gmsh/2018/011986.html

I consider this commit to be a bugfix, because with the OpenCascade
implementation inheriting from the builtin one in pygmsh, adding ellipsoids
was so far supported in the interface, but it yielded weird geometries
that were not even convex.

If you are interested, I will happily add a cutting mechanism similar
to the OC-Ball one to the Ellipsoid class.